change: Remove Commons CLI argument parsing
This beast of a system has been hard to deal with and optimize, and is not something I expect to be taken full advantage of. It's better for the developer that bloat like this is not included.
This commit also removes an unused 'raw' parameter in CommandData.

If you really need Commons CLI, see commit ddd575b. No support will be provided.