Relicense to GNU GPLv3
Initium is dropping the custom permissive licensed used in favor of the more free GPL license.
Any commit prior to this one is still licensed under the ZSLv1r0, however absolutely no support will be provided.

In the spirit of free software, I (Reperak), have decided to relicense Initium in the hopes that this will protect the freedom of Discord bot users. The majority of large Discord bots are closed source, meaning that Discord users are not only having their freedom revoked, but their privacy as well.

The GNU GPLv3 still permits private use, meaning there's no requirement to make your code public- just that if you do decide to, you must retain this license and grant the same freedoms that we (Zorroware) have granted to you.